The right version of get_next_line

+++ get_next_line/get_next_line_utils.c
@@ -6,7 +6,7 @@
 /*   By: bszikora <bszikora@student.42heilbronn.    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/03/27 16:27:00 by bszikora          #+#    #+#             */
-/*   Updated: 2024/04/15 17:10:52 by bszikora         ###   ########.fr       */
+/*   Updated: 2024/04/20 15:34:31 by bszikora         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -74,9 +74,9 @@ char	*ft_strchr(const char *s, int c)
 	return (0);
 }
 
-int	ft_strlen(const char *str)
+size_t	ft_strlen(const char *str)
 {
-	int	c;
+	size_t	c;
 
 	c = 0;
 	while (str[c] != '\0')
